Obtain path from the existing location state
`usePath` is a hook which also depends on `usePageLocation` and thus
adds another eventlistener on "locationchanged" and potentially
re-render when the state changes. As we already have `cockpit.location`
in state re-use this which saves setting up extra eventlisteners and
possibly re-renders.
